fix and add some steps

+++ mule-migration-tool-library/src/main/java/com/obi/tools/migration/library/smartgate/steps/properties/UpdateAutodicoveryStageProperties.java
@@ -36,7 +36,7 @@ public class UpdateAutodicoveryStageProperties implements NamespaceContribution 
   private static final String API_AUTODISCOVERY_NAME = "api.autodiscovery.name";
   private static final String API_AUTODISCOVERY_VERSION = "api.autodiscovery.version";
   private static final List<String> SMARGATE_STAGE_PROD_FILE_NAME =
-      Arrays.asList("prod.properties", "fqa.properties", "fqa.properties", "devtest.properties");
+      Arrays.asList("prod.properties", "fqa.properties", "fqa.properties", "devtest.properties", "local.properties");
   private static final String SMARTGATE_FILE_PATH = "src" + File.separator + "main" + File.separator + "resources";
 
   @Override
@@ -56,35 +56,36 @@ public class UpdateAutodicoveryStageProperties implements NamespaceContribution 
     try {
 
       File file = new File(appModel.getProjectBasePath().toFile(), SMARTGATE_FILE_PATH + File.separator + stageFileName);
-      PropertiesConfiguration config = new PropertiesConfiguration();
-      PropertiesConfigurationLayout layout = new PropertiesConfigurationLayout();
-      config.setLayout(layout);
-      layout.load(config, new InputStreamReader(new FileInputStream(file)));
+      if (file.exists()) {
+        PropertiesConfiguration config = new PropertiesConfiguration();
+        PropertiesConfigurationLayout layout = new PropertiesConfigurationLayout();
+        config.setLayout(layout);
+        layout.load(config, new InputStreamReader(new FileInputStream(file)));
 
-      // Search and modify API_AUTODISCOVERY_VERSION
-      String version = (String) config.getProperty(API_AUTODISCOVERY_VERSION);
-      if (version != null) {
-        version = version.replaceAll("v1:", "");
-        config.setProperty(API_AUTODISCOVERY_VERSION, version);
+        // Search and modify API_AUTODISCOVERY_VERSION
+        String version = (String) config.getProperty(API_AUTODISCOVERY_VERSION);
+        if (version != null) {
+          version = version.replaceAll("v1:", "");
+          config.setProperty(API_AUTODISCOVERY_VERSION, version);
+        }
+
+        // Search and modify API_AUTODISCOVERY_NAME
+        String name = (String) config.getProperty(API_AUTODISCOVERY_NAME);
+        if (name != null) {
+          Optional<PomModel> pomModel =
+              appModel.getPomModel();
+          config.setProperty(API_AUTODISCOVERY_NAME, pomModel.get().getArtifactId());
+        }
+
+        // Remove if exits
+        Object console = config.getProperty(API_AUTODISCOVERY_NAME);
+        if (console != null) {
+          config.clearProperty(API_CONSOLE_ENABLED);
+        }
+
+
+        layout.save(config, new FileWriter(file, false));
       }
-
-      // Search and modify API_AUTODISCOVERY_NAME
-      String name = (String) config.getProperty(API_AUTODISCOVERY_NAME);
-      if (name != null) {
-        Optional<PomModel> pomModel =
-            appModel.getPomModel();
-        config.setProperty(API_AUTODISCOVERY_NAME, pomModel.get().getArtifactId());
-      }
-
-      // Remove if exits
-      Object console = config.getProperty(API_AUTODISCOVERY_NAME);
-      if (console != null) {
-        config.clearProperty(API_CONSOLE_ENABLED);
-      }
-
-
-      layout.save(config, new FileWriter(file, false));
-
     } catch (IOException | ConfigurationException e) {
       throw new RuntimeException("Unable to modify File: " + stageFileName, e);
     }
